+++ CHANGELOG.md
@@ -3,9 +3,12 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## 0.5.0 - 2026-06-04
 ### Added
 - Active scan rule "PTK Scan Rules" that runs all of the PTK scan rules enabled.
+
+### Changed
+- Updated OWASP Penetration Testing Kit browser extension to 9.9.7
 
 ## 0.4.0
 ### Added